Move Constraint validator test case to Test namespace
