+++ src/Symfony/Component/Validator/Tests/Constraints/BicValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\BicValidator;
 use Symfony\Component\Validator\Constraints\Bic;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class BicValidatorTest extends AbstractConstraintValidatorTest
+class BicValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
