+++ src/Symfony/Component/Validator/Tests/Constraints/BlankValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\Blank;
 use Symfony\Component\Validator\Constraints\BlankValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class BlankValidatorTest extends AbstractConstraintValidatorTest
+class BlankValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
