+++ src/Symfony/Component/Validator/Tests/Constraints/CardSchemeValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\CardScheme;
 use Symfony\Component\Validator\Constraints\CardSchemeValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class CardSchemeValidatorTest extends AbstractConstraintValidatorTest
+class CardSchemeValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
