+++ src/Symfony/Component/Validator/Tests/Constraints/CollectionValidatorTest.php
@@ -17,8 +17,9 @@ use Symfony\Component\Validator\Constraints\NotNull;
 use Symfony\Component\Validator\Constraints\Optional;
 use Symfony\Component\Validator\Constraints\Range;
 use Symfony\Component\Validator\Constraints\Required;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-abstract class CollectionValidatorTest extends AbstractConstraintValidatorTest
+abstract class CollectionValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
