+++ src/Symfony/Component/Validator/Tests/Constraints/CountValidatorTest.php
@@ -13,11 +13,12 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\Count;
 use Symfony\Component\Validator\Constraints\CountValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
 /**
  * @author Bernhard Schussek <bschussek@gmail.com>
  */
-abstract class CountValidatorTest extends AbstractConstraintValidatorTest
+abstract class CountValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
