+++ src/Symfony/Component/Validator/Tests/Constraints/CurrencyValidatorTest.php
@@ -14,8 +14,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 use Symfony\Component\Intl\Util\IntlTestHelper;
 use Symfony\Component\Validator\Constraints\Currency;
 use Symfony\Component\Validator\Constraints\CurrencyValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class CurrencyValidatorTest extends AbstractConstraintValidatorTest
+class CurrencyValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
