+++ src/Symfony/Component/Validator/Tests/Constraints/DateValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\Date;
 use Symfony\Component\Validator\Constraints\DateValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class DateValidatorTest extends AbstractConstraintValidatorTest
+class DateValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
