+++ src/Symfony/Component/Validator/Tests/Constraints/IbanValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\Iban;
 use Symfony\Component\Validator\Constraints\IbanValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class IbanValidatorTest extends AbstractConstraintValidatorTest
+class IbanValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
