+++ src/Symfony/Component/Validator/Tests/Constraints/IpValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\Ip;
 use Symfony\Component\Validator\Constraints\IpValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class IpValidatorTest extends AbstractConstraintValidatorTest
+class IpValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
