+++ src/Symfony/Component/Validator/Tests/Constraints/IsFalseValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\IsFalse;
 use Symfony\Component\Validator\Constraints\IsFalseValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class IsFalseValidatorTest extends AbstractConstraintValidatorTest
+class IsFalseValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
