+++ src/Symfony/Component/Validator/Tests/Constraints/IsNullValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\IsNull;
 use Symfony\Component\Validator\Constraints\IsNullValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class IsNullValidatorTest extends AbstractConstraintValidatorTest
+class IsNullValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
