+++ src/Symfony/Component/Validator/Tests/Constraints/IsTrueValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\IsTrue;
 use Symfony\Component\Validator\Constraints\IsTrueValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class IsTrueValidatorTest extends AbstractConstraintValidatorTest
+class IsTrueValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
