+++ src/Symfony/Component/Validator/Tests/Constraints/IssnValidatorTest.php
@@ -13,11 +13,12 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\Issn;
 use Symfony\Component\Validator\Constraints\IssnValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
 /**
  * @see https://en.wikipedia.org/wiki/Issn
  */
-class IssnValidatorTest extends AbstractConstraintValidatorTest
+class IssnValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
