+++ src/Symfony/Component/Validator/Tests/Constraints/LuhnValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\Luhn;
 use Symfony\Component\Validator\Constraints\LuhnValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class LuhnValidatorTest extends AbstractConstraintValidatorTest
+class LuhnValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
