+++ src/Symfony/Component/Validator/Tests/Constraints/NotNullValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\NotNull;
 use Symfony\Component\Validator\Constraints\NotNullValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class NotNullValidatorTest extends AbstractConstraintValidatorTest
+class NotNullValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
