+++ src/Symfony/Component/Validator/Tests/Constraints/RangeValidatorTest.php
@@ -14,8 +14,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 use Symfony\Component\Intl\Util\IntlTestHelper;
 use Symfony\Component\Validator\Constraints\Range;
 use Symfony\Component\Validator\Constraints\RangeValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class RangeValidatorTest extends AbstractConstraintValidatorTest
+class RangeValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
