+++ src/Symfony/Component/Validator/Tests/Constraints/TypeValidatorTest.php
@@ -13,8 +13,9 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\Type;
 use Symfony\Component\Validator\Constraints\TypeValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class TypeValidatorTest extends AbstractConstraintValidatorTest
+class TypeValidatorTest extends ConstraintValidatorTestCase
 {
     protected static $file;
