+++ src/Symfony/Component/Validator/Tests/Constraints/UuidValidatorTest.php
@@ -13,11 +13,12 @@ namespace Symfony\Component\Validator\Tests\Constraints;
 
 use Symfony\Component\Validator\Constraints\Uuid;
 use Symfony\Component\Validator\Constraints\UuidValidator;
+use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
 /**
  * @author Colin O'Dell <colinodell@gmail.com>
  */
-class UuidValidatorTest extends AbstractConstraintValidatorTest
+class UuidValidatorTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
